optimize EachLocalQuad for complicated lattices
Problem: CCS wants to cares only about hopping directions, but for complex lattices these might depend on the basis.

Solution: Add `BondDirections` and `uc2diridx` to get dir indices for bonds based on the basis.

Note: This might be a slight performance loss for normal lattices but can be huge for complex lattices. (35% in test case)